Added the low level methods findAllAsResponse and createAsResponse.
We now have the low­level implementation of this CREATE API. We are, of course, doing the given authenticated just because CREATE and all of the operations that the API exposes all required authentication, so we are doingevenauthenticated here. We are specifying the content type this time because we are sending data. We are marshalling the role to the body here, and then we are simply posting to the main role's API, which is the collection resource, and we are posting in that resource.