Added higher level create method in the client.
We have a full implementation of this CREATE operation, which is now at the much higher level. It no longer returns a response. It actually goes back tothe API, retrieves the new role, and returns that new role. So let's have a look at the implementation. First of all, we are ofcourse doing the CREATE, and we are operating on the response from the API. We are then extracting this location header out of the response. This is what we expect the API will return. So basically, the API gives us all the information we need to discover the newly created resource, and it does so by returning the location of that resource via this location header