Copied RoleSimpleApiClient to GenericSimpleApiClient then abstracted the client and basically turning this role specific client into a client that is generic and that has no exact concept of the role resource.
Made this class abstract and this is because we're going to have to have some abstract method in here and we're going to extend this class whenever we are actually creating a client specifically for a single resource. We then add it to this generic parameter and we're going to basically be using these parameters throughout the implementation pointing to the specific type of that resource and we started to use this generic parameter here in the class and then passing in the class and so on.