Extended IDto because it has the getId method that is needed.
A compilation problem occurred here, resource get ID and this is because the generic parameter that we use has no ID or has no get ID implementation. And so we need to go a little bit more specific with that parameter... Now, luckily all of our DTO's implement this IDTO interface which actually has the withID here which does implement getID and setID. So luckily that is the case and now we can actually extend IDTO from this generic parameter and now, everything compiles correctly.