Make a simple api test into a contract test: start by changing the Role DTO into a String type.
The question now becomes, what does it take to start making this a contract test? What do we need to do in order to have a contract test and not just a simple API test?  01:35: Well, the very first thing that we'll need to do here is we'll need to stop operating on DTOs. We have here the resource and this is the actual DTO in our system. This is the role DTO. Now, instead of using that, we'll need to no longer rely on this particular DTO so we'll need to start adapting and using raw strings.