Externalized JSON data of the role into it's own role, but there is still hardcoding to deal with.
However, it's really important to understandthat this is just an intermediary step. This is not the final step, because if we were to run this test again, we will of course see a failure. And the reason for that is this name, which is supposed to be unique. And up until this point, it was unique. This is no longer unique. So we are going to have to make it unique without relying on the advantage of using DTOs, and we're going to do that next. But again, it's important to understand where we are. And where we are is that we started to externalize this JSON data of the role into its own file.