Make the name unique without relying on the advantage of using DTO
Well, the first thing we'll need to do is we'll need to parse out this raw JSON string into a tree representation, and of course we'll use Jackson for that. So pretty straightforward up to this point. We're simply instantiating a new object mapper, and we are reading this string into the tree representation. And now we have the root node to work with. The next step is going to be editing the name node and changing that to a random value. We had a hard­coded value, and now we need to change that to something random.

First, we needed to cast this JsonNode into an object node. This will give us some extra API methods, and one of those methods, one of the methods that we wanted to get access to by doing the cast is this set API. So we can see that this set API actually sits in this object node implementation, so that's why we needed to cast. Now, via this set API, we are simply editing the name node, and we are essentially replacing that with a new text node. And so this text node that has a random string replaces the old name node. And now that the replacement is done, allwe need to do is to return the root node asstring. And that is it. We should now have a fully working test that instead of failing after the first pass, it should keep passing no matter how manytimes we run it.