Added swagger dependency, configured Swagger in UmWebConfig
We defined the main configuration here. This is the docket bean. This is going to be our entry point in controlling the entire configuration of Swagger.

Use the builder pattern in order to further configure this docket, and thefirst thing we did is that we ensure that Swagger will pick up all of our request handlers that define our Spring API. We did that with this predicate requesthandlerselectors.any(). And as you can see, there are a few options here. We want ini because we only have the API in this particular application. But if your application is bundling more than just an API, then you can certainly go in and select a more restrictive predicate here and basically pinpoint the API that you need covered by Swagger.

We are doing the exact same thing for our paths because basically, all of them need to be covered by the Swagger docs. Next, we're specifying to Swagger what the base path of our servlet, the servlet that is serving our API, what that is, and in our case, it's /API.

And now, we're doing some extra configuration. This is a very straightforward way to configure global substitutions of models, which in this case basically means that we want local dates to simply be replaced by strings.

This is definitely an interesting configuration aspect, is that we really want Swagger to not document response entities. What we actually want is the data within the response entity. So the response entity wraps a certain kind of data and uses that certain kind of data as a generic parameter. And so we don't want the response the entity, which is a Spring­specific type of data. We don't want that showing up in our Swagger documentation. We want the actual data type that gets wrapped in this response entity. That's what we really want to show up in our main documentation. And of course in that way the documentation remains highly relevant as well as Spring agnostic. We don't really want to Spring specific artefacts such as the response entity, we don't want that in our main Swagger documentation. So there we have it. This is basically the main Swagger documentation, and of course this is something that we can keep tuning and keep configuring, but this is certainly a very good base to build on because Swagger really doesn't need that much configuration.